frontend-is-ready; separated pages
