+++ src/pages/rescueTheTux/teamRegistration/success.js
@@ -1,6 +1,6 @@
 import TerminalWindow from '../../../components/terminal/terminalWindow';
 
-export default function Success( {nextStep, prevStep, loginTeamPage} ) {
+export default function Success( {teamValues, loginTeamPage} ) {
 		
 	return (
 		<TerminalWindow
@@ -12,11 +12,10 @@ export default function Success( {nextStep, prevStep, loginTeamPage} ) {
 		>
 		<div className='form-start'>Success!...</div>
 		<div>
-		     Dear, Thank you for joining the team to rescue our Tux! See you soon on 19th August! :)
+		     Dear {teamValues.gname}, Thank you for joining the team to rescue our Tux! See you soon on 19th August! :)
 	        </div>
 		<div className='form-end'>
 		<button className='form-nav-button next' onClick={loginTeamPage}> Login </button>
-	        <button className='form-nav-button' onClick={prevStep}>Back </button>
 		</div>
 		</TerminalWindow>
 	);
